Added column to db, created profile api controller. Moved changedesc api in it.
