+++ src/frontend/services/user-profile-service.js
@@ -1,0 +1,9 @@
+import { HttpService } from "./http-service";
+
+export class UserProfileService {
+  static async setUserDescription(newDescription) {
+    await HttpService.httpPost(
+      `/api/profile/changedesc?newDesc=${newDescription}`
+    );
+  }
+}
